requirements.txt
###### Requirements without Version Specifiers ######
numpy
streamlit
pillow
torch
torchvision